backend/requirements.txt
django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.1
python-decouple==3.8
fastapi==0.121.1
uvicorn==0.38.0
httpx==0.28.1
python-multipart==0.0.6
pydantic>=2.7.0
google-generativeai>=0.3.0
dj-database-url>=1.0.0
whitenoise>=6.5.0
psycopg2-binary>=2.9.7
gunicorn>=20.1.0
redis>=4.5.0
rq>=1.13.0
prometheus-client>=0.16.0